change weight and price fields from int to float

+++ src/domain/dish.go
@@ -10,8 +10,8 @@ import (
 type Dish struct {
 	Base
 	Name       string       `json:"name" gorm:"not null" binding:"required"`
-	Weight     int          `json:"weight" gorm:"not null" binding:"required"`
-	Price      int          `json:"price" gorm:"not null" binding:"required"`
+	Weight     float32      `json:"weight" gorm:"not null" binding:"required"`
+	Price      float32      `json:"price" gorm:"not null" binding:"required"`
 	Desc       string       `json:"desc"`
 	Images     []ImageArray `json:"images"`
 	CateringID uuid.UUID    `json:"-"`
